+++ src/test/java/com/swagfish/server/WebServerTest.java
@@ -1,11 +1,11 @@
-package com.swagfish.tictactoe;
+package com.swagfish.server;
 
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.hamcrest.CoreMatchers.containsString;
 import org.junit.Test;
 
-public class MainTest 
+public class WebServerTest 
 {
 	@Test
 	public void test()
